Change type of port for setSerialPort to Stream
This allows devs to use any kind of Stream, for example SoftwareSerial,
to connect to the VESC. Since that works just fine, it's not necessarry
to limit it to just HardwareSerial.
Note: I don't know if there are any scenarios where someone would want
to use anything except for a Serial port, but maybe there is some weird
case where it'd make sense to bridge the communication between Arduino
and VESC via some other communication protocol, so I went with Stream
instead of Serial.